Rename cops_api module to corgi_api
